vmselect: simplify errors returned to user for instant and range handlers
Internal function `Exec()` is used for evaluating expressions received by
`instant` or `range` handlers. Under the hood, `Exec` may be recursively
called while traversing expression parts. When `Exec` fails with the error,
the error stack contains all the errors for each `part` of the expression
which was recursivelly called. This results into a very complex and hard
to read errors.

The change introduces a new error type `ExecErr` for errors returned by
`Exec` fn. This error type is then checked in the HTTP handler in order
to strip off all the context except the last error message. This supposed
to improve error readability in web interfaces such as VMUI or Grafana.

Consider the following example.
Before:
```
422
error when executing query="(node_filesystem_avail_bytes * 100) / node_filesystem_size_bytes < 10 and ON (instance, device, mountpoint) node_filesystem_readonly == 0" on the time range (start=1658142907487, end=1658315707487, step=540000):
    cannot execute query:
        cannot execute "(((node_filesystem_avail_bytes * 100) / node_filesystem_size_bytes) < 10) and on (instance, device, mountpoint) (node_filesystem_readonly == 0)":
            cannot execute "node_filesystem_readonly == 0":
                cannot evaluate "node_filesystem_readonly":
                    search error after reading 0 data blocks:
                        error when searching for tagFilters=[{__name__="node_filesystem_readonly"}] on the time range [2022-07-18T11:01:07.487Z..2022-07-20T11:15:07.487Z]:
                            error when searching tsids:
                                the number of matching timeseries exceeds 1; either narrow down the search or increase -search.max* command-line flag values at vmselect; see https://docs.victoriametrics.com/#resource-usage-limits
```

After:
```
422
the number of matching timeseries exceeds 1; either narrow down the search or increase -search.max* command-line flag values at vmselect; see https://docs.victoriametrics.com/#resource-usage-limits
```

In the `After` section we don't have all the initial context. But the idea
is that user does not need that context on the client side, so user receives
only the last error which is the exact reason why query failed.

The full error trace is still logged with the full context
and can be found in vmselect logs.

Signed-off-by: hagen1778 <roman@victoriametrics.com>